Fix the failure test for unknown months.
I previously changed the underlying datatype from a list to a dict, to allow for multiple abbreviated forms of a month. Due to the corresponding changes in the iteration code, it meant that the failure scenario of an unknown month presented in a different manner. I’ve updated the test to reflect the behaviour of unknown months.