Cleanup S3 and image pipelines
The S3 pipeline now appears to use the v4 signature by default so there
is no need to maintain a specific class to do that. Similarly the Scrapy
image pipeline can handle saving to S3 allowing for large cleanup in the
ImageDownloader class as well.

This introduces a change for the fields available in the JobsItem so it
may be necessary to update tests in the sqlalchemy-mappings project.